move webhook router to a non-internal cmd package so that other places can import it to register routes
