+++ cmd/frontend/webhooks/github_webhooks.go
@@ -1,4 +1,4 @@
-package httpapi
+package webhooks
 
 import (
 	"context"
@@ -57,6 +57,7 @@ func (h *GithubWebhook) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
+	log15.Warn(eventType)
 	// match event handlers
 	err = h.Dispatch(r.Context(), eventType, extSvc, e)
 	if err != nil {
