+++ cmd/frontend/webhooks/github_webhooks_test.go
@@ -1,4 +1,4 @@
-package httpapi
+package webhooks
 
 import (
 	"bytes"
